Struct trapframe::arch::UserContext

#[repr(C)]
pub struct UserContext {
    pub tls: usize,
    pub __reserved: usize,
    pub status: usize,
    pub cause: usize,
    pub epc: usize,
    pub vaddr: usize,
    pub general: GeneralRegs,
}

Saved registers on a trap.

Fields
tls: usize

TLS

__reserved: usize

Reserved for internal use

status: usize

CP0 Status

cause: usize

CP0 cause

epc: usize

CP0 epc

vaddr: usize

CP0 vaddr

general: GeneralRegs

General registers

Implementations

impl UserContext

pub fn run(&mut self)

Go to user space with the context, and come back when a trap occurs.

On return, the context will be reset to the status before the trap. Trap reason and error code will be returned.

Example

use trapframe::{UserContext, GeneralRegs};

// init user space context
let mut context = UserContext {
    general: GeneralRegs {
        sp: 0x10000,
        ..Default::default()
    },
    sepc: 0x1000,
    ..Default::default()
};
// go to user
context.run();
// back from user
println!("back from user: {:#x?}", context);

impl UserContext

pub fn get_syscall_num(&self) -> usize

Get number of syscall

pub fn get_syscall_ret(&self) -> usize

Get return value of syscall

pub fn set_syscall_ret(&mut self, ret: usize)

Set return value of syscall

pub fn get_syscall_args(&self) -> [usize; 6]

Get syscall args

pub fn set_ip(&mut self, ip: usize)

Set instruction pointer

pub fn set_sp(&mut self, sp: usize)

Set stack pointer

pub fn get_sp(&self) -> usize

Get stack pointer

pub fn set_tls(&mut self, tls: usize)

Set tls pointer