Struct trapframe::UserContext

#[repr(C)]
pub struct UserContext {
    pub general: GeneralRegs,
    pub trap_num: usize,
    pub error_code: usize,
}

User space context

Fields
general: GeneralRegstrap_num: usizeerror_code: usize
Implementations

impl UserContext

pub fn run_fncall(&mut self)

Go to user context by function return, within the same privilege level.

User program should call syscall_fn_entry() to return back. Trap reason and error code will always be set to 0x100 and 0.

impl UserContext

pub fn get_syscall_num(&self) -> usize

Get number of syscall

pub fn get_syscall_ret(&self) -> usize

Get return value of syscall

pub fn set_syscall_ret(&mut self, ret: usize)

Set return value of syscall

pub fn get_syscall_args(&self) -> [usize; 6]

Get syscall args

pub fn set_ip(&mut self, ip: usize)

Set instruction pointer

pub fn set_sp(&mut self, sp: usize)

Set stack pointer

pub fn get_sp(&self) -> usize

Get stack pointer

pub fn set_tls(&mut self, tls: usize)

Set tls pointer